Refactor tests, remove unneeded config
