--- .cargo/config
@@ -1,2 +1,0 @@
-[net]
-git-fetch-with-cli = true
