+++ src/client.rs
@@ -109,31 +109,17 @@ fn deserialize_response(mut raw_resp: Vec<u8>) -> Result<Response<Vec<u8>>> {
 
 #[cfg(test)]
 mod tests {
-    use tempdir::TempDir;
-    use std::fs::File;
-    use std::io::Write;
+    use crate::tests;
 
     use super::*;
 
     #[test]
     fn test_get() {
         crate::tests::setup_tracing();
-
-        let tempdir = TempDir::new("tor-cache").expect("create temp dir");
-
-        let consensus = include_str!("test-data/consensus.txt");
-        let microdescriptors = include_str!("test-data/microdescriptors.txt");
-
-        let consensus_path = tempdir.path().join("consensus.txt");
-        let mut consensus_file = File::create(consensus_path).expect("create temp consensus file");
-        write!(consensus_file, "{}", consensus).expect("write temp consensus file");
-
-        let microdescriptors_path = tempdir.path().join("microdescriptors.txt");
-        let mut microdescriptors_file = File::create(microdescriptors_path).expect("create temp microdescriptors file");
-        write!(microdescriptors_file, "{}", microdescriptors).expect("write temp microdescriptors file");
+        let docdir = tests::setup_docdir();
 
         let resp = Client::new(DirectoryCache {
-            tmp_dir: tempdir.path().to_str().map(String::from),
+            tmp_dir: docdir.path().to_str().map(String::from),
             nodes: None,
             relays: None,
         })
